:panda_face:     add   global option

+++ crates/command/src/reset.rs
@@ -1,13 +1,17 @@
-use crate::init_hyperscoop;
+use crate::init_env::{get_apps_path, get_apps_path_global};
 use crate::utils::utility::compare_versions;
 use anyhow::bail;
+use crossterm::style::Stylize;
 use std::cmp::Ordering;
 use std::os::windows::fs::symlink_dir;
 
-pub fn reset_latest_version(name: String) -> Result<(), anyhow::Error> {
-    let hyperscoop = init_hyperscoop()?;
-    let app_name = hyperscoop.get_apps_path();
-    for entry in std::fs::read_dir(app_name)? {
+pub fn reset_latest_version(name: String, global: bool) -> Result<(), anyhow::Error> {
+    let app_root_dir = if global {
+        get_apps_path_global()
+    } else {
+        get_apps_path()
+    };
+    for entry in std::fs::read_dir(app_root_dir)? {
         let entry = entry?;
         let path = entry.path();
         if path.is_dir() {
@@ -25,19 +29,30 @@ pub fn reset_latest_version(name: String) -> Result<(), anyhow::Error> {
             } else if count == 2 {
                 let current_path = path.join("current");
                 log::info!("Resetting app: {}", current_path.display());
-                std::fs::remove_dir_all(&current_path).expect("Failed to remove current directory");
+                if !current_path.exists() {
+                    bail!("Current directory is not exist");
+                }
+                std::fs::remove_dir_all(&current_path)?;
                 for entry in std::fs::read_dir(&path)? {
                     let entry = entry?;
                     let path = entry.path();
                     if !path.is_dir() {
                         continue;
                     }
-                    symlink_dir(path, current_path.as_path())
+                    symlink_dir(&path, current_path.as_path())
                         .expect("Failed to create app symlink");
+                    println!(
+                        "Resetting app  successfully {}=>{}",
+                        current_path.display().to_string().dark_green().bold(),
+                        path.display().to_string().dark_green().bold()
+                    );
                 }
                 return Ok(());
             } else {
                 let current_path = path.join("current");
+                if !current_path.exists() {
+                    bail!("Current directory is not exist");
+                };
                 std::fs::remove_dir_all(&current_path).unwrap_or_else(|e| {
                     eprintln!("Failed to remove current directory, error: {}", e);
                 });
@@ -60,20 +75,28 @@ pub fn reset_latest_version(name: String) -> Result<(), anyhow::Error> {
                 let max_version_path = path.join(&max_version);
                 log::info!("Resetting app: {}", max_version_path.display());
                 symlink_dir(max_version_path, current_path.as_path())
-                    .expect("Failed to create app symlink"); 
+                    .expect("Failed to create app symlink");
                 println!("Reset {}@{} success", name, &max_version);
                 return Ok(());
             }
         }
     }
-    bail!("App not found: {}", name)  
+    bail!("App not found: {}", name)
 }
 
-pub fn reset_specific_version(name: String, version: String) -> Result<(), anyhow::Error> {
+pub fn reset_specific_version(
+    name: &str,
+    version: &str,
+    global: bool,
+) -> Result<(), anyhow::Error> {
     log::info!("Resetting app: {}@{}", name, version);
-    let hyperscoop = init_hyperscoop()?;
-    let app_name = hyperscoop.get_apps_path();
-    for entry in std::fs::read_dir(app_name)? {
+    let app_root_dir = if global {
+        get_apps_path_global()
+    } else {
+        get_apps_path()
+    };
+
+    for entry in std::fs::read_dir(app_root_dir)? {
         let entry = entry?;
         let path = entry.path();
         if path.is_dir() {
@@ -88,21 +111,21 @@ pub fn reset_specific_version(name: String, version: String) -> Result<(), anyho
             }
             if count <= 1 {
                 bail!("app 文件目录格式不正确,不存在Current目录或者至少一个版本目录")
-            } else { 
+            } else {
                 let current_path = path.join("current");
-                
-                let version_path = path.join(&version); 
-               if !version_path.exists() {
-                   bail!("Version not found: {}", version );
-               }
-              std::fs::remove_dir_all(&current_path).unwrap_or_else(|e| {
-                eprintln!("Failed to remove current directory, error: {}", e);
-              });
-              symlink_dir(version_path, current_path.as_path())?  ; 
-              println!("Reset {}@{} success", name, &version);  
-              return Ok(());
+
+                let version_path = path.join(&version);
+                if !version_path.exists() {
+                    bail!("Version dir {} is not exist ", version);
+                }
+                std::fs::remove_dir_all(&current_path).unwrap_or_else(|e| {
+                    eprintln!("Failed to remove current directory, error: {}", e);
+                });
+                symlink_dir(version_path, current_path.as_path())?;
+                println!("Reset {}@{} success", name, &version);
+                return Ok(());
             }
         }
     }
-    bail!("App not found: {}", name);  
+    bail!("App not found: {}", name);
 }
